cpg.h, coroapi.h: more size_t and const changes
* include/corosync/cpg.h (cpg_deliver_fn_t, cpg_confchg_fn_t)
(cpg_groups_get_fn_t):
* include/corosync/engine/coroapi.h (ipc_response_send)
(ipc_dispatch_send, tpg_join, tpg_leave, tpg_groups_mcast)
(tpg_groups_reserve):

git-svn-id: http://svn.fedorahosted.org/svn/corosync/trunk@2043 fd59a12c-fef9-0310-b244-a6a79926bd2f

+++ include/corosync/engine/coroapi.h
@@ -392,12 +392,12 @@ struct corosync_api_v1 {
 
 	void *(*ipc_private_data_get) (void *conn);
 
-	int (*ipc_response_send) (void *conn, const void *msg, int mlen);
+	int (*ipc_response_send) (void *conn, const void *msg, size_t mlen);
 
 	int (*ipc_response_iov_send) (void *conn,
 				      const struct iovec *iov, unsigned int iov_len);
 
-	int (*ipc_dispatch_send) (void *conn, const void *msg, int mlen);
+	int (*ipc_dispatch_send) (void *conn, const void *msg, size_t mlen);
 
 	int (*ipc_dispatch_iov_send) (void *conn,
 				      const struct iovec *iov, unsigned int iov_len);
@@ -465,12 +465,12 @@ struct corosync_api_v1 {
 	int (*tpg_join) (
 		hdb_handle_t handle,
 		struct corosync_tpg_group *groups,
-		int group_cnt);
+		size_t group_cnt);
 
 	int (*tpg_leave) (
 		hdb_handle_t handle,
 		struct corosync_tpg_group *groups,
-		int group_cnt);
+		size_t group_cnt);
 
 	int (*tpg_joined_mcast) (
 		hdb_handle_t handle,
@@ -490,14 +490,14 @@ struct corosync_api_v1 {
 		hdb_handle_t handle,
 		int guarantee,
 		const struct corosync_tpg_group *groups,
-		int groups_cnt,
+		size_t groups_cnt,
 		const struct iovec *iovec,
 		unsigned int iov_len);
 
 	int (*tpg_groups_reserve) (
 		hdb_handle_t handle,
 		const struct corosync_tpg_group *groups,
-		int groups_cnt,
+		size_t groups_cnt,
 		const struct iovec *iovec,
 		unsigned int iov_len);
